+++ fastapi_mvc/__init__.py
@@ -1,8 +1,8 @@
-# -*- coding: utf-8 -*-
 """fastapi-mvc."""
 import logging
 
 from .version import __version__  # noqa: F401
+
 
 # initialize logging
 log = logging.getLogger(__name__)