+++ fastapi_mvc/app/exceptions/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""fastapi-mvc."""
-from .http import HTTPException, http_exception_handler
-
-
-__all__ = (HTTPException, http_exception_handler)