+++ fastapi_mvc/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/app/utils/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""{{cookiecutter.project_description}}"""
+from .aiohttp_client import AiohttpClient
+from .redis import RedisClient
+
+__all__ = (
+    AiohttpClient,
+    RedisClient,
+)