+++ fastapi_mvc/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/cli/cli.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+"""{{cookiecutter.project_name}} CLI root."""
+import logging
+
+import click
+from {{cookiecutter.package_name}}.cli.commands.serve import serve
+
+
+@click.group()
+@click.option(
+    "-v",
+    "--verbose",
+    help="Enable verbose logging.",
+    is_flag=True,
+    default=False,
+)
+def cli(**options):
+    """{{cookiecutter.project_name}} CLI root."""
+    if options["verbose"]:
+        level = logging.DEBUG
+    else:
+        level = logging.INFO
+
+    logging.basicConfig(
+        level=level,
+        format="[%(asctime)s] [%(process)s] [%(levelname)s] %(message)s",
+        datefmt="%Y-%m-%d %H:%M:%S %z",
+    )
+
+
+cli.add_command(serve)