+++ fastapi_mvc/template/{{cookiecutter.folder_name}}/{{cookiecutter.package_name}}/config/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""{{cookiecutter.project_description}}"""
+from .application import settings
+from .redis import redis
+from .router import router
+
+
+__all__ = (
+    settings,
+    redis,
+    router,
+)