+++ tests/unit/cli/test_cli.py
@@ -7,9 +7,9 @@ def test_root(cli_runner):
 
 
 def test_root_options(cli_runner):
-    result = cli_runner.invoke(cli, ["--verbose", "serve", "--help"])
+    result = cli_runner.invoke(cli, ["--verbose", "new", "--help"])
     assert result.exit_code == 0
-    result = cli_runner.invoke(cli, ["serve", "--help"])
+    result = cli_runner.invoke(cli, ["new", "--help"])
     assert result.exit_code == 0